+++ media_kit/lib/media_kit.dart
@@ -6,21 +6,16 @@
 
 // ignore_for_file: camel_case_types
 
-export 'package:media_kit/src/player.dart';
 export 'package:media_kit/src/media_kit.dart';
-export 'package:media_kit/src/platform_player.dart' hide Player;
+export 'package:media_kit/src/player/player.dart';
+export 'package:media_kit/src/player/platform_player.dart';
 
-export 'package:media_kit/src/models/media/media.dart';
+export 'package:media_kit/src/player/libmpv/player/player.dart';
+export 'package:media_kit/src/player/web/player/player.dart';
+
 export 'package:media_kit/src/models/track.dart';
 export 'package:media_kit/src/models/playlist.dart';
+export 'package:media_kit/src/models/media/media.dart';
 export 'package:media_kit/src/models/audio_device.dart';
 export 'package:media_kit/src/models/audio_params.dart';
 export 'package:media_kit/src/models/playlist_mode.dart';
-
-// For invoking platform specific API.
-
-import 'package:media_kit/src/libmpv/player.dart'
-    if (dart.library.html) 'package:media_kit/src/platform_player.dart'
-    as _libmpv show Player;
-
-typedef libmpvPlayer = _libmpv.Player;
