+++ media_kit/lib/src/models/media/media_native.dart
@@ -11,7 +11,7 @@ import 'package:uri_parser/uri_parser.dart';
 import 'package:safe_local_storage/safe_local_storage.dart';
 
 import 'package:media_kit/src/models/playable.dart';
-import 'package:media_kit/src/android_asset_loader.dart';
+import 'package:media_kit/src/utils/android_asset_loader.dart';
 
 HashMap<String, Media> medias = HashMap<String, Media>();
 HashMap<String, double> bitrates = HashMap<String, double>();
