+++ media_kit/lib/src/player/libmpv/core/fallback_bitrate_handler.dart
