+++ media_kit/lib/src/player/libmpv/player/stub.dart
@@ -3,11 +3,11 @@
 /// Copyright © 2021 & onwards, Hitesh Kumar Saini <saini123hitesh@gmail.com>.
 /// All rights reserved.
 /// Use of this source code is governed by MIT license that can be found in the LICENSE file.
+// ignore_for_file: camel_case_types
+import 'package:media_kit/src/player/platform_player.dart';
 
-/// Initializes the libmpv backend for package:media_kit.
-///
-/// Following optional parameters are available:
-/// * `libmpv`: Manually specified the path to the libmpv shared library.
-void ensureInitialized({String? libmpv}) {
-  // N/A
+void libmpvEnsureInitialized({String? libmpv}) {}
+
+class libmpvPlayer extends PlatformPlayer {
+  libmpvPlayer({required super.configuration});
 }
