+++ media_kit/lib/src/player/web/player/player.dart
@@ -1,0 +1,7 @@
+/// This file is a part of media_kit (https://github.com/alexmercerind/media_kit).
+///
+/// Copyright © 2021 & onwards, Hitesh Kumar Saini <saini123hitesh@gmail.com>.
+/// All rights reserved.
+/// Use of this source code is governed by MIT license that can be found in the LICENSE file.
+
+export 'real.dart' if (dart.library.io) 'stub.dart';
