+++ media_kit/lib/src/utils/web.dart
