+++ media_kit/test/src/player/libmpv/core/native_library_test.dart
@@ -6,7 +6,7 @@
 
 import 'package:test/test.dart';
 
-import 'package:media_kit/src/libmpv/core/native_library.dart';
+import 'package:media_kit/src/player/libmpv/core/native_library.dart';
 
 void main() {
   test(
